Add a who-am-I endpoint
It's extremely useful for serverless single page application who want to know if the user is logged in and what's their email.

There's already an endpoint /showUser that is not exactly doing what it says, it simply returns `You are logged in` but has some iframe logic in it, so it didn't feel right to change it.

`/whoami` returns the JSON representation of the user or 401 if not logged in, which is what an API should do