Use Composer profile to reach the pan-domain-auth-settings bucket.
The pan-domain-auth-settings bucket has a policy which allows each of the Tools accounts to GetObject. This means we can use Composer credentials to access the required files to validate a cookie. That is, we can simplify the requirements of this app to a single set of Janus credentials, rather than two.