Update code owner to workflow-and-collaboration
The [Workflow and Collaboration
team](https://github.com/orgs/guardian/teams/workflow-and-collaboration)
are [taking ownership of
panda](https://github.com/guardian/pan-domain-authentication/pull/189),
and we feel it makes sense for us to also take ownership of this
repository as it is so closely related. This commit therefore sets the
code owner for this repository to be us.