Ensure devs can easily use the required version of Java
In a recent rehearsal of the Emergency access system:

https://docs.google.com/document/d/1NPoaY0xxO2SI6G3k9y7ov_GXPdMbS9iM8WppIP-z2BA/edit?usp=sharing

...we encountered this error which trying to run `sbt console`:

```
[info] compiling 32 Scala sources and 1 Java source to /Users/Roberto_Tyley/code/login.gutools/target/scala-2.13/classes ...
[error] /Users/Roberto_Tyley/code/login.gutools/app/config/LoginConfig.scala:72:17: constructor URL in class URL is deprecated (since 20)
[error]       val url = new URL(returnUrl)
[error]                 ^
[error] one error found
[error] (Compile / compileIncremental) Compilation failed
```

The https://github.com/guardian/login.gutools repo did not have a `.tool-versions` file,
and so I was running with an arbitrary choice of the Java 21 LTS. However, Java 20
deprecated `java.net.URL`, and `login.gutools` was configured to treat deprecation as fatal:

https://github.com/guardian/login.gutools/blob/5e1608206931928dcd1684a709b58b8e77ecf6a6/build.sbt#L8-L10

This fix includes `.tool-versions` file that makes explicit that this project should now
build with Java 21, and updates the CI configuration so that `setup-scala` will take advantage
of the `.tool-versions` file and use the _same_ version of Java as the developer. The AMI
image used has also been updated to use Java 21.