Reinstate the 1-minute refresh, to avoid indirect reliance on the healthcheck
Sam Hessian pointed out, while reviewing at commit d15940072, that there's a flaw
in the argument I made for completely removing the explicit 1-minute schedule that
reads the switches value:

https://github.com/guardian/login.gutools/pull/100#discussion_r2231499125

...Sam pointed out that promptly notifying when the switch is altered in S3 is critical
functionality for login.gutools, and with this PR, absent any other human-initiated
requests to read Switches, the only route for notifying when S3 changes are the
requests made by the healthcheck - and that places an unintuitive critical role on the
healthcheck. Healthchecks play an obvious role in checking the health of an application,
but it would be not immediately obvious to a dev that the healthcheck is also critical
to ensuring that switch-changes are notified.

Therefore, I'm reinstating the 1-minute schedule to read the switches on a regular basis.